backend/requirements.txt
aiofiles
annotated-types
anyio
bcrypt
certifi
charset-normalizer
click
fastapi
h11
idna
PyJWT
pydantic-settings
numpy
pandas
mysql-connector-python
pydantic
pydantic[email]
pydantic-settings
pydantic_core
python-dateutil
python-dotenv
pytz
requests
six
sniffio
starlette
typing-inspection
typing_extensions
tzdata
urllib3
uvicorn
